moving vault actions to new file
