+++ blockchain/calls/proxyActions/adapters/DssProxyActionsSmartContractWrapperInterface.ts
@@ -3,7 +3,38 @@ import {
   PayableTransactionObject,
 } from '../../../../types/web3-v1-contracts/types'
 import { ContextConnected } from '../../../network'
-import { DepositAndGenerateData, OpenData, WithdrawAndPaybackData } from '../proxyActions'
+import { TxMetaKind } from '../../txMeta'
+import { BigNumber } from 'bignumber.js'
+
+export type WithdrawAndPaybackData = {
+  kind: TxMetaKind.withdrawAndPayback
+  id: BigNumber
+  token: string
+  ilk: string
+  withdrawAmount: BigNumber
+  paybackAmount: BigNumber
+  proxyAddress: string
+  shouldPaybackAll: boolean
+}
+
+export type DepositAndGenerateData = {
+  kind: TxMetaKind.depositAndGenerate
+  id: BigNumber
+  token: string
+  ilk: string
+  depositAmount: BigNumber
+  generateAmount: BigNumber
+  proxyAddress: string
+}
+
+export type OpenData = {
+  kind: TxMetaKind.open
+  token: string
+  ilk: string
+  depositAmount: BigNumber
+  generateAmount: BigNumber
+  proxyAddress: string
+}
 
 export interface DssProxyActionsSmartContractWrapperInterface {
   resolveContractAddress: (context: ContextConnected) => string
