+++ blockchain/calls/proxyActions/adapters/curveDssProxyActionsSmartContractWrapper.ts
@@ -3,8 +3,12 @@ import {
   PayableTransactionObject,
 } from '../../../../types/web3-v1-contracts/types'
 import { ContextConnected } from '../../../network'
-import { DssProxyActionsSmartContractWrapperInterface } from './DssProxyActionsSmartContractWrapperInterface'
-import { DepositAndGenerateData, OpenData, WithdrawAndPaybackData } from '../proxyActions'
+import {
+  DepositAndGenerateData,
+  DssProxyActionsSmartContractWrapperInterface,
+  OpenData,
+  WithdrawAndPaybackData,
+} from './DssProxyActionsSmartContractWrapperInterface'
 
 export const CurveDssProxyActionsSmartContractWrapper: DssProxyActionsSmartContractWrapperInterface = {
   open(context: ContextConnected, data: OpenData): NonPayableTransactionObject<string> {
