+++ blockchain/calls/proxyActions/adapters/dssProxyActionsSmartContractWrapper.mock.ts
@@ -4,8 +4,12 @@ import {
   PayableTx,
 } from '../../../../types/web3-v1-contracts/types'
 import { ContextConnected } from '../../../network'
-import { DssProxyActionsSmartContractWrapperInterface } from './DssProxyActionsSmartContractWrapperInterface'
-import { DepositAndGenerateData, OpenData, WithdrawAndPaybackData } from '../proxyActions'
+import {
+  DepositAndGenerateData,
+  DssProxyActionsSmartContractWrapperInterface,
+  OpenData,
+  WithdrawAndPaybackData,
+} from './DssProxyActionsSmartContractWrapperInterface'
 
 function createMockPayableTransactionObject<T>(name: string): PayableTransactionObject<T> {
   return {
