+++ features/borrow/manage/pipes/viewStateTransforms/manageVaultTransitions.ts
@@ -3,7 +3,6 @@ import { TxHelpers } from 'components/AppContext'
 import { zero } from 'helpers/zero'
 import { Observable } from 'rxjs'
 
-import { VaultActionsLogicInterface } from '../../../../../blockchain/calls/proxyActions/proxyActions'
 import {
   defaultMutableManageVaultState,
   ManageStandardBorrowVaultState,
@@ -18,6 +17,7 @@ import {
   manageVaultDepositAndGenerate,
   manageVaultWithdrawAndPayback,
 } from './manageVaultTransactions'
+import { VaultActionsLogicInterface } from '../../../../../blockchain/calls/proxyActions/vaultActionsLogicFactory'
 
 type ManageVaultMultiplyTransitionChange =
   | {
